requirements.txt
pandas
openpyxl
python-dotenv
qrcode[pil]